refactor: Dict and Idx API (#27)
Refactors the Dict and Idx APIs so that the respective data structures have ownership over their readers.
Fixes #26

+++ internal/testutil/idx.go
@@ -23,7 +23,7 @@ import (
 )
 
 // MakeIndex make a test index given a list of words.
-func MakeIndex(words []*idx.Word, idxoffsetbits int64) []byte {
+func MakeIndex(words []*idx.Word, idxoffsetbits int) []byte {
 	b := []byte{}
 	for _, w := range words {
 		b = append(b, []byte(w.Word)...)
